added confirmation check for the user

+++ picturegallery/src/main/java/picturegallery/action/SearchIdenticalAction.java
@@ -49,6 +49,13 @@ public class SearchIdenticalAction extends Action {
 				nextState.registerAction(Action.createTempAction(KeyCode.D, "delete duplicated identical pictures", new ActionRunnable() {
 					@Override
 					public void run(State currentState) {
+						// TODO: eigentlich müsste vorher zu State geschlossen werden, damit keine (nicht mehr vorhandenen!!) Pictures nachgeladen werden können!!
+						// the key will not be deleted => one picture will be kept!!
+						if (!Logic.askForConfirmation("Delete duplicated items?", "In the collection " + selection
+								+ ", there are " + result.size() + " pictures with duplicates!",
+								"Do you want to delete all the duplicated pictures?")) {
+							return;
+						}
 						for (Entry<Picture, List<Picture>> e : result.entrySet()) {
 							for (Picture picToDelete : e.getValue()) {
 								MainApp.get().deletePicture(picToDelete, false);
